Refactor objectType store
