+++ src/client/components/Navigation/Topnav.js
@@ -23,7 +23,7 @@ import {
 import listOfObjectTypes from '../../../common/constants/listOfObjectTypes';
 import { replacer } from '../../helpers/parser';
 import { getObjectName } from '../../helpers/wObjectHelper';
-import { setFiltersAndLoad } from '../../objectTypes/objectTypeActions';
+import { setFiltersAndLoad } from '../../store/objectTypeStore/objectTypeActions';
 import HeaderButton from '../HeaderButton/HeaderButton';
 import { getTranformSearchCountData, pendingSearch } from '../../search/helpers';
 import UserSearchItem from '../../search/SearchItems/UserSearchItem';
