+++ src/client/discoverObjects/DiscoverObjectsContent.js
@@ -11,19 +11,7 @@ import {
   updateActiveFilters,
   updateActiveTagsFilters,
 } from './helper';
-import {
-  getActiveFilters,
-  getObjectTypeSorting,
-  getObjectTypeState,
-  getObjectTypeLoading,
-  getFilteredObjects,
-  getHasMoreRelatedObjects,
-  getAvailableFilters,
-  getHasMap,
-  getIsMapModalOpen,
-  getFiltersTags,
-  getActiveFiltersTags,
-} from '../store/reducers';
+import { getIsMapModalOpen } from '../store/reducers';
 import {
   getObjectTypeByStateFilters,
   clearType,
@@ -33,7 +21,7 @@ import {
   setActiveFilters,
   setTagsFiltersAndLoad,
   setActiveTagsFilters,
-} from '../objectTypes/objectTypeActions';
+} from '../store/objectTypeStore/objectTypeActions';
 import { setMapFullscreenMode } from '../components/Maps/mapActions';
 import Loading from '../components/Icon/Loading';
 import ObjectCardView from '../objectCard/ObjectCardView';
@@ -50,6 +38,18 @@ import { RADIUS, ZOOM } from '../../common/constants/map';
 import { getCryptoPriceHistory } from '../store/appStore/appActions';
 import { HBD, HIVE } from '../../common/constants/cryptos';
 import { getAuthenticatedUserName } from '../store/authStore/authSelectors';
+import {
+  getActiveFilters,
+  getActiveFiltersTags,
+  getAvailableFilters,
+  getFilteredObjects,
+  getFiltersTags,
+  getHasMap,
+  getHasMoreRelatedObjects,
+  getObjectTypeLoading,
+  getObjectTypeSorting,
+  getObjectTypeState,
+} from '../store/objectTypeStore/objectTypeSelectors';
 
 const modalName = {
   FILTERS: 'filters',
