+++ src/client/store/objectTypeStore/objectTypeActions.js
@@ -1,16 +1,14 @@
-import { createAsyncActionType } from '../helpers/stateHelpers';
+import { createAsyncActionType } from '../../helpers/stateHelpers';
+import { getUserLocation, getQueryString, getLocale } from '../reducers';
+import * as ApiClient from '../../../waivioApi/ApiClient';
+import { createFilterBody } from '../../discoverObjects/helper';
+import { getAuthenticatedUserName } from '../authStore/authSelectors';
 import {
   getActiveFilters,
-  getTypeName,
-  getObjectTypeSorting,
-  getUserLocation,
-  getQueryString,
-  getLocale,
   getActiveFiltersTags,
-} from '../store/reducers';
-import * as ApiClient from '../../waivioApi/ApiClient';
-import { createFilterBody } from '../discoverObjects/helper';
-import { getAuthenticatedUserName } from '../store/authStore/authSelectors';
+  getObjectTypeSorting,
+  getTypeName,
+} from './objectTypeSelectors';
 
 export const GET_OBJECT_TYPE = createAsyncActionType('@objectType/GET_OBJECT_TYPE');
 export const GET_OBJECT_TYPE_MAP = createAsyncActionType('@objectType/GET_OBJECT_TYPE_MAP');
